Checkpoint: initial moving things around
