+++ app/App.tsx
@@ -3,7 +3,6 @@ import React, { useEffect, useState } from "react";
 // package and all references to it:
 // https://support.google.com/analytics/answer/12938611#zippy=%2Cin-this-article
 import ReactGA_4 from "react-ga4";
-import Intercom from "react-intercom";
 import { Helmet } from "react-helmet-async";
 import { useHistory } from "react-router-dom";
 import {
@@ -12,18 +11,113 @@ import {
   websiteConfig,
   AppProvider,
 } from "./utils";
-import { Footer, Navigation, UserWay } from "./components/ui";
+import { Footer, Navigation } from "./components/ui";
 
 import config from "./config";
 import MetaImage from "./assets/img/Our415_OG.png";
 import styles from "./App.module.scss";
+import { Configure, InstantSearch } from "react-instantsearch-core";
+import { liteClient } from "algoliasearch/lite";
+import { useLocation } from "react-router-dom";
+import { useCookies } from "react-cookie";
+import qs, { ParsedQs } from "qs";
 
-const { intercom, siteUrl, title, userWay } = websiteConfig;
+import { DEFAULT_AROUND_PRECISION, useAppContext, websiteConfig } from "utils";
+import { translate } from "utils/DataService";
+
+import { Loader } from "components/ui";
+import SearchResults from "components/search/SearchResults/SearchResults";
+import Sidebar from "components/search/Sidebar/Sidebar";
+import { Header } from "components/search/Header/Header";
+import { SiteSearchInput } from "components/ui/SiteSearchInput";
+import { AroundRadius } from "algoliasearch";
+import config from "./config";
+import styles from "./SearchResultsPage.module.scss";
+
+const { siteUrl, title } = websiteConfig;
 export const OUTER_CONTAINER_ID = "outer-container";
+const searchClient = liteClient(
+  config.ALGOLIA_APPLICATION_ID,
+  config.ALGOLIA_READ_ONLY_API_KEY
+);
+
+interface ConfigureState {
+  aroundRadius?: AroundRadius;
+  [key: string]: any;
+}
+
+const INDEX_NAME = `${config.ALGOLIA_INDEX_PREFIX}_services_search`;
+
+interface SearchState extends ParsedQs {
+  configure?: ConfigureState;
+  query?: string | null;
+  [key: string]: any;
+}
 
 export const App = () => {
+  const [cookies] = useCookies(["googtrans"]);
+  const { search } = useLocation();
   const history = useHistory();
   const [userLocation, setUserLocation] = useState<GeoCoordinates | null>(null);
+  const [searchState, setSearchState] = useState<SearchState | null>(null);
+  // const [aroundLatLang, setAroundLatLng] = useState({
+  //   lat: userLocation?.lat,
+  //   lng: userLocation?.lng,
+  // });
+  // const [searchRadius, setSearchRadius] = useState(
+  //   searchState?.configure?.aroundRadius ?? "all"
+  // );
+  // In cases where we translate a query into English, we use this value
+  // to represent the user's original, untranslated input. The untranslatedQuery
+  // is displayed in the UI and stored in the URL params.
+  // This untranslatedQuery value is also checked against when a new search is triggered
+  // to determine if the user has input a different query (vs. merely selecting refinements),
+  // in which case we need to call the translation API again
+  const [untranslatedQuery, setUntranslatedQuery] = useState<string | null>(
+    null
+  );
+  const [translatedQuery, setTranslatedQuery] = useState<string | null>(null);
+  const [nonQuerySearchParams, setNonQuerySearchParams] = useState<SearchState>(
+    {}
+  );
+
+  useEffect(() => {
+    const qsParams = qs.parse(search.slice(1));
+    setUntranslatedQuery(
+      qsParams[`${INDEX_NAME}.query}`]
+        ? (qsParams[`${INDEX_NAME}.query}`] as string)
+        : ""
+    );
+    delete qsParams["production_services_search[query]"];
+    setNonQuerySearchParams(qsParams);
+  }, [search]);
+
+  useEffect(() => {
+    if (untranslatedQuery === null) {
+      return;
+    }
+
+    let queryLanguage = "en";
+    // Google Translate determines translation source and target with a
+    // "googtrans" cookie. If the cookie exists, we assume that the
+    // the query should be translated into English prior to querying Algolia
+    const translationCookie = cookies.googtrans;
+    if (translationCookie) {
+      [, queryLanguage] = translationCookie.split("/en/");
+    }
+
+    const emptyQuery = untranslatedQuery.length === 0;
+    if (queryLanguage === "en" || emptyQuery) {
+      setTranslatedQuery(untranslatedQuery);
+    } else if (untranslatedQuery) {
+      translate(untranslatedQuery, queryLanguage).then((result) =>
+        setTranslatedQuery(result)
+      );
+    }
+  }, [untranslatedQuery, cookies.googtrans]);
+  useEffect(() => {
+    setSearchState({ ...nonQuerySearchParams, query: translatedQuery });
+  }, [translatedQuery, nonQuerySearchParams]);
 
   useEffect(() => {
     getLocation().then((loc) => {
@@ -48,6 +142,10 @@ export const App = () => {
     });
   }, [history]);
 
+  if (userLocation === null) {
+    return <Loader />;
+  }
+
   return (
     <div id={OUTER_CONTAINER_ID} className={styles.outerContainer}>
       <AppProvider userLocation={userLocation}>
@@ -68,11 +166,18 @@ export const App = () => {
           <meta property="og:image:width" content="1200" />
           <meta property="og:image:height" content="630" />
         </Helmet>
-        {userWay && <UserWay appID={config.SFFAMILIES_USERWAY_APP_ID} />}
-        {intercom && config.INTERCOM_APP_ID && (
-          <Intercom appID={config.INTERCOM_APP_ID} />
-        )}
-        <Navigation />
+        <InstantSearch
+          searchClient={searchClient}
+          indexName={INDEX_NAME}
+          routing
+        >
+          <Configure
+            aroundLatLng={`${userLocation.aroundLatLang.lat}, ${userLocation.aroundLatLang.lng}`}
+            aroundRadius={"all"}
+            aroundPrecision={DEFAULT_AROUND_PRECISION}
+          />
+          <Navigation />
+        </InstantSearch>
         <Footer />
       </AppProvider>
     </div>
