+++ app/config.ts
@@ -12,12 +12,9 @@ interface Config {
   AUTH0_REDIRECT_URI: string;
   GOOGLE_ANALYTICS_GA4_ID: string;
   GOOGLE_API_KEY: string;
-  INTERCOM_APP_ID: string;
   LINKSF_DOMAIN: string;
   MOHCD_DOMAIN: string;
   MOHCD_SUBDOMAIN: string;
-  SFFAMILIES_DOMAIN: string;
-  SFFAMILIES_USERWAY_APP_ID: string;
   STRAPI_API_TOKEN: string;
   STRAPI_API_URL: string;
 }
@@ -32,12 +29,9 @@ const config: Config = {
   AUTH0_REDIRECT_URI: CONFIG.AUTH0_REDIRECT_URI,
   GOOGLE_ANALYTICS_GA4_ID: process.env.NODE_ENV === "production" ? "" : "",
   GOOGLE_API_KEY: CONFIG.GOOGLE_API_KEY,
-  INTERCOM_APP_ID: "",
   LINKSF_DOMAIN: CONFIG.LINKSF_DOMAIN,
   MOHCD_DOMAIN: CONFIG.MOHCD_DOMAIN,
   MOHCD_SUBDOMAIN: CONFIG.MOHCD_SUBDOMAIN,
-  SFFAMILIES_DOMAIN: CONFIG.SFFAMILIES_DOMAIN,
-  SFFAMILIES_USERWAY_APP_ID: "4EhNmEIwPz",
   STRAPI_API_TOKEN: CONFIG.STRAPI_API_TOKEN,
   STRAPI_API_URL: CONFIG.STRAPI_API_URL,
 };
