+++ app/styles/utils/_helpers.scss
@@ -164,15 +164,6 @@ p.message {
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////
-// INTERCOM OVERRIDES
-.intercom-launcher-frame {
-  z-index: 3 !important;
-  @media screen and (max-width: 800px) {
-    display: none !important;
-  }
-}
-
-////////////////////////////////////////////////////////////////////////////////////////
 // Typography Classes
 
 .heading-1 {
