+++ app/utils/useAppContext.tsx
@@ -3,10 +3,16 @@ import { GeoCoordinates } from "utils";
 
 interface Context {
   userLocation: GeoCoordinates | null;
+  aroundUserLocationRadius?: "all" | number;
+}
+
+interface AppProviderProps extends Context {
+  children: React.ReactNode;
 }
 
 export const AppContext = createContext<Context>({
   userLocation: null,
+  aroundUserLocationRadius: "all",
 });
 
 export const useAppContext = () => useContext(AppContext);
@@ -14,10 +20,8 @@ export const useAppContext = () => useContext(AppContext);
 export const AppProvider = ({
   children,
   userLocation,
-}: {
-  children: React.ReactNode;
-  userLocation: GeoCoordinates | null;
-}) => {
+  aroundUserLocationRadius,
+}: AppProviderProps) => {
   // We have to use useMemo here to manage the contextValue to ensure that the user's authState
   // propagates downward after authentication. I couldn't find a way to get this to work with
   // useState. Moreover, we can't use a simple object to define contextValue, as the object would
@@ -25,10 +29,13 @@ export const AppProvider = ({
   const contextValue = useMemo(() => {
     return {
       userLocation,
+      aroundUserLocationRadius,
     };
-  }, [userLocation]);
+  }, [userLocation, aroundUserLocationRadius]);
 
   return (
     <AppContext.Provider value={contextValue}>{children}</AppContext.Provider>
   );
 };
+
+// NEEd to add setting function in ehre
