+++ package.json
@@ -41,7 +41,6 @@
     "react-helmet-async": "^1.3.0",
     "react-instantsearch": "^7.13.0",
     "react-instantsearch-core": "^7.13.0",
-    "react-intercom": "^1.0.15",
     "react-markdown": "^8.0.7",
     "react-modal": "^3.12.1",
     "react-router-dom": "^5.1.2",
